PLG-116: Separating session functionality in new cartEvents helper class
