+++ Block/Analytics.php
@@ -77,13 +77,13 @@ class Analytics extends Template
         return in_array($action, $rejected);
     }
     
-    public function getCatalogSession() {
-        return $this->helper->catalogSession;
+    public function getCartEvents() {
+        return $this->helper->cartEvents;
     }
     
     public function cartEventsLoop($eventsData, $eventMethod){
         foreach($eventsData as $event) {
-            echo "metrilo." . $eventMethod . "('" . $event['data']['productId'] . "', " . $event['data']['quantity'] . ");";
+            echo "metrilo." . $eventMethod . "('" . $event['productId'] . "', " . $event['quantity'] . ");";
         }
     }
 }
